requirements.txt
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
DateTime==4.4
dbus-python==1.2.16
Deprecated==1.2.13
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
gitdb==4.0.9
GitPython==3.1.27
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lizard==1.17.10
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.22.3
oauthlib==3.1.0
pandas==1.4.2
pexpect==4.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
PyDriller==2.1
PyGithub==1.55
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.1
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
smmap==5.0.0
sos==4.0
ssh-import-id==5.10
systemd-python==234
Twisted==18.9.0
types-pytz==2021.3.8
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wrapt==1.14.1
zipp==1.0.0
zope.interface==4.7.1